+++ test/nextgen/operators/matmul/matmul/matmul_ukerapi_wrapper.hpp
@@ -9,13 +9,12 @@
 #include <string>
 #include <string_view>
 
-#include "kai/ukernels/matmul/kai_matmul.h"
 #include "kai/ukernels/matmul/kai_matmul_types.h"
 #include "test/nextgen/common/poly.hpp"
 #include "test/nextgen/format/format.hpp"
 #include "test/nextgen/harness/kernel_wrapper.hpp"
-#include "test/nextgen/operators/matmul/matmul/matmul_interface.hpp"
 #include "test/nextgen/operators/matmul/matmul_dims.hpp"
+#include "test/nextgen/operators/matmul/matmul_slots.hpp"
 
 namespace kai::test {
 
@@ -24,11 +23,12 @@ class MatMulUkerApiWrapper : public KernelWrapper<MatMulShape> {
 public:
     /// Creates a new wrapper.
     MatMulUkerApiWrapper(
-        std::string_view name, const Poly<Format>& lhs_format, const Poly<Format>& rhs_format,
-        const Poly<Format>& dst_format) :
+        std::string_view name, kai_matmul_uker_api api, MatMulSlot lhs_input_slot, const Poly<Format>& lhs_format,
+        const Poly<Format>& rhs_format, const Poly<Format>& dst_format) :
         m_name(name),
         m_uker_config(),
-        m_ukernel(kai_matmul_clamp_f32_f32p4vsx1_f32p4vsx1bf32_8vsx8vs_sme2_mopa()),
+        m_ukernel(api),
+        m_lhs_input_slot(lhs_input_slot),
         m_lhs_format(lhs_format),
         m_rhs_format(rhs_format),
         m_dst_format(dst_format) {
@@ -47,6 +47,7 @@ private:
     std::string m_name;
     kai_matmul_uker_config m_uker_config;
     kai_matmul_uker_api m_ukernel;
+    MatMulSlot m_lhs_input_slot;
     Poly<Format> m_lhs_format;
     Poly<Format> m_rhs_format;
     Poly<Format> m_dst_format;